sorting and one object for all entries
i created a common.js that should have the functions that all entries
need to use. One of those is the Entry object the only weird one in
there could be the detail and that is what it will say before the date.
example: "in theaters: " I also made a comparer so that when we put
things together they will be date sorted. and finally i made a
createEntry which will add the given entry to the results div. this is
so that everything looks the same among different tabs.