Banner!
Canvas banner works, randomly selects images from the current page and
displays them in the banner.  We need to find a place where "drawBanner"
can be called so that it is called only after the page is finished
loading.

Changed CSS to have black border around white text
(-webkit-text-stroke...)  Will need to implement this cross-browser.